Refactor opening of external editor windows
