+++ newIDE/app/src/ObjectEditor/Editors/SpriteEditor/SpritesList.js
@@ -25,7 +25,6 @@ import { type ResourceExternalEditor } from '../../../ResourcesList/ResourceExte
 import { applyResourceDefaults } from '../../../ResourcesList/ResourceUtils';
 import RaisedButtonWithSplitMenu from '../../../UI/RaisedButtonWithSplitMenu';
 const gd: libGDevelop = global.gd;
-const path = require('path');
 
 const SPRITE_SIZE = 100; //TODO: Factor with Thumbnail
